requirements.txt
# 核心依赖
openai>=1.0.0              # LLM客户端（兼容千问/DeepSeek）
gradio>=4.0.0              # Web界面
python-dotenv>=1.0.0       # 环境变量管理

# 数据库
# SQLite 内置，无需额外依赖

# 开发依赖
pytest>=7.0.0              # 测试框架
pytest-asyncio>=0.21.0     # 异步测试支持

